services/newsletter-engine/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
google-generativeai>=0.4.0
pydantic>=2.6.0
python-dotenv>=1.0.1
